+++ lib/src/presentation/widgets/questionnaire_item/choice/questionnaire_radio_button_choice_item_view.dart
@@ -30,7 +30,7 @@ class QuestionnaireRadioButtonChoiceItemViewState
                 title: Text(valueNameResolver(entry)),
                 value: entry,
                 groupValue: selectedValue,
-                onChanged: onSelectedValueChanged,
+                onChanged: isReadOnly ? null : onSelectedValueChanged,
               ))
           .toList(),
     );