+++ lib/src/presentation/widgets/questionnaire_item/base/questionnaire_multi_choice_item_view.dart
@@ -1,4 +1,3 @@
-import 'package:collection/collection.dart';
 import 'package:fhir/r4.dart';
 import 'package:fhir_questionnaire/fhir_questionnaire.dart';
 import 'package:flutter/material.dart';
@@ -30,13 +29,21 @@ abstract class QuestionnaireMultiChoiceItemViewState<
   void initState() {
     super.initState();
     if (controller.value.isEmpty) {
-      final initial =
-          item.initial?.firstWhereOrNull((item) => item.valueCoding != null);
+      List<QuestionnaireAnswerOption> initial = [];
+      for (final value in item.initial ?? <QuestionnaireInitial>[]) {
+        if (value.valueCoding != null) {
+          initial.add(QuestionnaireAnswerOption(
+            valueCoding: value.valueCoding!,
+          ));
+        } else if (value.valueString != null) {
+          initial.add(QuestionnaireAnswerOption(
+            valueString: value.valueString,
+          ));
+        }
+      }
 
-      if (initial?.valueCoding != null) {
-        selectedValues.add(QuestionnaireAnswerOption(
-          valueCoding: initial?.valueCoding!,
-        ));
+      if (initial.isNotEmpty) {
+        selectedValues.addAll(initial);
       }
     }
   }