+++ lib/src/presentation/widgets/size_renderer.dart
@@ -25,7 +25,7 @@ class _SizeRendererState extends State<SizeRenderer> {
     SchedulerBinding.instance.addPostFrameCallback(onCreated);
   }
 
-  void onCreated(_) {
+  void onCreated(Duration _) {
     var context = key.currentContext;
     if (context == null) return;